+++ sdk/core/core-rest-pipeline/src/constants.ts
@@ -4,3 +4,7 @@
 export const SDK_VERSION: string = "1.10.0";
 
 export const DEFAULT_RETRY_POLICY_COUNT = 3;
+
+// intervals are in milliseconds
+export const DEFAULT_CLIENT_RETRY_INTERVAL = 1000;
+export const DEFAULT_CLIENT_MAX_RETRY_INTERVAL = 1000 * 64;