+++ sdk/core/core-rest-pipeline/src/interfaces.ts
@@ -3,6 +3,7 @@
 
 import { AbortSignalLike } from "@azure/abort-controller";
 import { OperationTracingOptions } from "@azure/core-tracing";
+import { FailoverHostGenerator } from "./retryStrategies/failoverRetryStrategy";
 
 /**
  * A HttpHeaders collection represented as a simple JSON object.
@@ -340,6 +341,11 @@ export interface PipelineRetryOptions {
    * to 64000 (64 seconds).
    */
   maxRetryDelayInMs?: number;
+
+  /**
+   * A generator for failover endpoints.
+   */
+  failoverHostGenerator?: FailoverHostGenerator;
 }
 
 /**