+++ sdk/tables/data-tables/src/TableClient.ts
@@ -42,7 +42,7 @@ import {
   isSASCredential,
   isTokenCredential,
 } from "@azure/core-auth";
-import { Pipeline, retryPolicy, throttlingRetryStrategy } from "@azure/core-rest-pipeline";
+import { defaultFailoverHostGenerator, Pipeline } from "@azure/core-rest-pipeline";
 import { STORAGE_SCOPE, TablesLoggingAllowedHeaderNames } from "./utils/constants";
 import { decodeContinuationToken, encodeContinuationToken } from "./utils/continuationToken";
 import {
@@ -64,7 +64,6 @@ import { Uuid } from "./utils/uuid";
 import { apiVersionPolicy } from "./utils/apiVersionPolicy";
 import { cosmosPatchPolicy } from "./cosmosPathPolicy";
 import { escapeQuotes } from "./odata";
-import { georedundantRetryStrategy } from "./policies/georedundantRetryStrategy";
 import { getClientParamsFromConnectionString } from "./utils/connectionString";
 import { handleTableAlreadyExists } from "./utils/errorHelpers";
 import { isCosmosEndpoint } from "./utils/isCosmosEndpoint";
@@ -250,6 +249,13 @@ export class TableClient {
       },
     };
 
+    if (this.clientOptions.secondaryEndpoint) {
+      internalPipelineOptions.retryOptions = {
+        ...internalPipelineOptions.retryOptions,
+        failoverHostGenerator: defaultFailoverHostGenerator([this.clientOptions.secondaryEndpoint]),
+      };
+    }
+
     const generatedClient = new GeneratedClient(this.url, internalPipelineOptions);
     if (isNamedKeyCredential(credential)) {
       generatedClient.pipeline.addPolicy(tablesNamedKeyCredentialPolicy(credential));
@@ -267,25 +273,6 @@ export class TableClient {
 
     if (options.version) {
       generatedClient.pipeline.addPolicy(apiVersionPolicy(options.version));
-    }
-
-    if (options.georedundantRetryOptions) {
-      const maxRetries =
-        (options.georedundantRetryOptions.maxRetries ?? 3) *
-        Math.max(
-          options.georedundantRetryOptions.readHosts?.length ?? 1,
-          options.georedundantRetryOptions.writeHosts?.length ?? 1
-        );
-      generatedClient.pipeline.removePolicy({ phase: "Retry" });
-      generatedClient.pipeline.addPolicy(
-        retryPolicy(
-          [throttlingRetryStrategy(), georedundantRetryStrategy(options.georedundantRetryOptions)],
-          {
-            maxRetries,
-          }
-        ),
-        { phase: "Retry" }
-      );
     }
 
     this.generatedClient = generatedClient;